Moved loggedOn sharedPref to welcome fragment n fixed a bug in cart fragment
