+++ app/src/main/java/com/decagon/avalanche/views/fragments/WelcomeFragment.kt
@@ -1,11 +1,14 @@
 package com.decagon.avalanche.views.fragments
 
 import android.app.ActivityOptions
+import android.content.SharedPreferences
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
 import com.decagon.avalanche.R
 import com.decagon.avalanche.databinding.FragmentWelcomeBinding
@@ -14,6 +17,15 @@ import com.decagon.avalanche.databinding.FragmentWelcomeBinding
 class WelcomeFragment : Fragment() {
     private var _binding: FragmentWelcomeBinding? = null
     private val binding get() = _binding!!
+    lateinit var loggedOnSharePref: SharedPreferences
+
+    override fun onStart() {
+        super.onStart()
+
+        /** Check onBoarding user status */
+        checkIfUserLoggedOn()
+    }
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -22,12 +34,6 @@ class WelcomeFragment : Fragment() {
         _binding = FragmentWelcomeBinding.inflate(inflater, container, false)
 
         binding.loginBtn.setOnClickListener {
-//            val pairs: Array<Pair<View, String>?> = arrayOfNulls(1)
-//            pairs[0] = Pair<View, String>(binding.loginBtn, "transition_login")
-//
-//            val options =
-//                ActivityOptions.makeSceneTransitionAnimation(requireActivity(), pairs)
-
             findNavController().navigate(R.id.loginFragment)
         }
 
@@ -35,6 +41,18 @@ class WelcomeFragment : Fragment() {
             findNavController().navigate(R.id.signUpFragment)
         }
         return binding.root
+    }
+
+    private fun checkIfUserLoggedOn() {
+
+        loggedOnSharePref = requireActivity().getSharedPreferences("loggedOn", AppCompatActivity.MODE_PRIVATE)
+        var isFirstTime: Boolean = loggedOnSharePref.getBoolean("firstTime", true)
+
+
+        /** Check if user has not logged on before, else move to login screen */
+        if (!isFirstTime) {
+            findNavController().navigate(R.id.loginFragment)
+        }
     }
